requirements.txt
# Core Data Processing
pandas>=2.1.0
numpy>=1.24.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# NLP & Text Processing
spacy>=3.7.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Vector Database
qdrant-client>=1.7.0

# LLM & AI
transformers>=4.35.0
torch>=2.1.0
bitsandbytes>=0.41.0
accelerate>=0.24.0

# Multi-Agent Framework
crewai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.1

# LLM Providers
groq>=0.4.0
google-generativeai>=0.4.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Configuration & Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0

# HTTP & APIs
httpx>=0.25.0
aiohttp>=3.9.0,<3.11.0
requests>=2.31.0
feedparser>=6.0.0

# Google Services (Location Agent)
googlemaps>=4.10.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
